fix to asr, addition of mov.b and mov.l to fix zero extends, expansion of the monitor code to be 32K x 32 instead of 16K

+++ monitor/include/integer.h
@@ -25,8 +25,8 @@ typedef int				INT;
 typedef unsigned int	UINT;
 
 /* These types MUST be 32 bit */
-typedef long			LONG;
-typedef unsigned long	DWORD;
+typedef int			LONG;
+typedef unsigned int	DWORD;
 
 #endif
